이미지 다중 저장된것 출력중

+++ src/main/java/net/daum/dao/PostRepository.java
@@ -1,11 +1,29 @@
 package net.daum.dao;
 
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
 import net.daum.vo.Community_boardVO;
 
-public interface PostRepository extends JpaRepository<Community_boardVO, Integer> {
+public interface PostRepository extends JpaRepository<Community_boardVO, Long> {
+
+	
+
+	
+
+
+
+
+
+
+
+	
+
+	
+
+
